create_bank.rebuild_mode: commit latest changes

+++ app/create_bank/_rebuild_mode.py
@@ -52,25 +52,27 @@ class RebuildMode:
         boot_dir: str,
         stats_tracker: OtaClientStatistics,
         status_updator: Callable,
+        old_bank_root: str,
     ) -> None:
         from app.proxy_info import proxy_cfg
 
         self.cookies = cookies
         self.metadata = metadata
         self.url_base = url_base
-        self.mount_point = Path(mount_point)
+        self.old_bank_root = Path(old_bank_root)  # TODO: old bank
+        self.new_bank_root = Path(mount_point)
         self.boot_dir = Path(boot_dir)
         self.stats_tracker = stats_tracker
         self.status_update: Callable = status_updator
-
         # the location of image at the ota server root
         self.image_base_dir = self.metadata.get_rootfsdir_info()["file"]
         self.image_base_url = urljoin(url_base, f"{self.image_base_dir}/")
 
         # temp storage
-        self._meta_storage = tempfile.TemporaryDirectory(prefix="ota_metadata")
-        self._meta_folder = Path(self._meta_storage.name)
-        self._tmp_store = tempfile.TemporaryDirectory(prefix="ota_tmpstore")
+        # TODO: considering cross reboot persistent of /tmp/store
+        # TODO: reused if update interrupted, unconditionally
+        self._meta_folder: Path = None  # TODO: configured by cfg
+        self._tmp_store = Path("/var/tmp/ota-tmp")  # TODO: configured by cfg
         self._tmp_folder = Path(self._tmp_store.name)
 
         # configure the downloader
@@ -79,10 +81,6 @@ class RebuildMode:
         if proxy:
             logger.info(f"use {proxy=} for downloading")
             self._downloader.configure_proxy(proxy)
-
-    def __del__(self):
-        self._meta_storage.cleanup()
-        self._tmp_store.cleanup()
 
     def _prepare_meta_files(self):
         for fname, method in self.META_FILES:
@@ -102,10 +100,15 @@ class RebuildMode:
         delta_calculator = DeltaGenerator(
             old_reg="/opt/ota/image_meta/regulars.txt",
             new_reg=self._meta_folder / "regulars.txt",
+            bank_root=self.old_bank_root,
         )
-        _new, _hold, _ = delta_calculator.calculate_delta()
-        self._new = _new
-        self._hold = _hold
+
+        (
+            self._new,
+            self._hold,
+            self._rm,
+            self.total_files_num,
+        ) = delta_calculator.calculate_delta()
 
     def _process_persistents(self):
         """NOTE: just copy from legacy mode"""
@@ -117,8 +120,8 @@ class RebuildMode:
         _copy_tree = CopyTree(
             src_passwd_file=_passwd_file,
             src_group_file=_group_file,
-            dst_passwd_file=self.mount_point / _passwd_file.relative_to("/"),
-            dst_group_file=self.mount_point / _group_file.relative_to("/"),
+            dst_passwd_file=self.new_bank_root / _passwd_file.relative_to("/"),
+            dst_group_file=self.new_bank_root / _group_file.relative_to("/"),
         )
 
         with open(self._meta_folder / "persistents.txt", "r") as f:
@@ -129,81 +132,85 @@ class RebuildMode:
                     or perinf.path.is_dir()
                     or perinf.path.is_symlink()
                 ):  # NOTE: not equivalent to perinf.path.exists()
-                    _copy_tree.copy_with_parents(perinf.path, self.mount_point)
+                    _copy_tree.copy_with_parents(perinf.path, self.new_bank_root)
 
     def _process_dirs(self):
         self.status_update(OtaClientUpdatePhase.DIRECTORY)
         with open(self._meta_folder / "dirs.txt", "r") as f:
             for l in f:
-                DirectoryInf(l).mkdir2dst(self.mount_point)
+                DirectoryInf(l).mkdir2bank(self.new_bank_root)
 
         # TODO: save metadata to /opt/ota/image_meta
 
     def _process_symlinks(self):
         with open(self._meta_folder / "symlinks.txt", "r") as f:
             for l in f:
-                SymbolicLinkInf(l).symlink2dst(self.mount_point)
+                SymbolicLinkInf(l).symlink2bank(self.new_bank_root)
 
     def _process_regulars(self):
         self.status_update(OtaClientUpdatePhase.REGULAR)
-        with open(self._meta_folder / "regulars.txt", "r") as f:
-            total_files_num = len(f.readlines())
 
-        self.stats_tracker.set("total_regular_files", total_files_num)
-
-        _hardlink_register = HardlinkRegister()
-        _download_se = Semaphore(self.MAX_CONCURRENT_DOWNLOAD)
+        self._hardlink_register = HardlinkRegister()
+        self._download_se = Semaphore(self.MAX_CONCURRENT_DOWNLOAD)
         _collector = CreateRegularStatsCollector(
             self.stats_tracker,
-            total_regular_num=total_files_num,
+            total_regular_num=self.total_files_num,
             max_concurrency_tasks=self.MAX_CONCURRENT_TASKS,
         )
 
         with ThreadPoolExecutor(thread_name_prefix="create_standby_bank") as pool:
-            # fire up background collector
-            pool.submit(_collector.collector)
+            # collect recycled files from _rm
+            futs = []
+            for _, _pathset in self._rm.items():
+                futs.append(
+                    pool.submit(
+                        _pathset.collect_entries_to_be_recycled,
+                        dst=self._tmp_folder,
+                        root=self.old_bank_root,
+                    )
+                )
 
+            concurrent_futures_wait(futs)
+            del futs  # cleanup
+
+            # apply delta _hold and _new
+            pool.submit(_collector.collector)
             for _hash, _regulars_set in itertools.chain(
                 self._hold.items(), self._new.items()
             ):
                 _collector.acquire_se()
                 fut = pool.submit(
-                    self._create_from_regs_set,
+                    self._apply_delta,
                     _hash,
                     _regulars_set,
-                    hardlink_register=_hardlink_register,
-                    download_se=_download_se,
                 )
                 fut.add_done_callback(_collector.callback)
 
-            logger.info("all create_regular_files tasks dispatched, wait for collector")
+            logger.info("all process_regulars tasks dispatched, wait for finishing")
             _collector.wait()
 
-    def _create_from_regs_set(
-        self,
-        _hash: str,
-        _regs_set: RegularInfSet,
-        *,
-        hardlink_register: HardlinkRegister,
-        download_se: Semaphore,
-    ) -> List[RegularStats]:
-        _first_copy: Path = self._tmp_folder / _hash
-        _first_copy_prepared = Event()
-        if _first_copy.is_file():
-            _first_copy_prepared.set()
+    def _apply_delta(self, _hash: str, _regs_set: RegularInfSet) -> List[RegularStats]:
+        stats_list = []
+        skip_cleanup = _regs_set.skip_cleanup
 
-        stats_list: List[RegularStats] = []
-        for is_last, entry in _regs_set:
+        _first_copy = self._tmp_folder / _hash
+        for is_last, entry in _regs_set.iter_entries():
             _start = time.thread_time()
             _stat = RegularStats()
 
             # prepare first copy for the hash group
-            if not _first_copy_prepared.is_set():
-                if _regs_set.is_first_copy_available() or (
-                    is_last and not entry.path.is_file()
-                ):  # download the file to the tmp dir
+            if not _first_copy.is_file():
+                _collected_entry = _regs_set.entry_to_be_collected
+                try:
+                    if _collected_entry is None:
+                        raise FileNotFoundError
+
+                    # copy from the current bank
+                    _collected_entry.copy2dst(_first_copy, src_root=self.old_bank_root)
+                    _stat.op = "copy"
+                except FileNotFoundError:  # fallback to download from remote
                     _stat.op = "download"
-                    with download_se:  # limit on-going downloading
+                    with self._download_se:  # limit on-going downloading
                         _stat.errors = self._downloader.download(
                             entry.path,
                             _first_copy,
@@ -211,52 +218,46 @@ class RebuildMode:
                             url_base=self.image_base_url,
                             cookies=self.cookies,
                         )
-                else:  # copy from current bank
-                    _stat.op = "copy"
-                    entry.copy2dst(self._tmp_folder / _hash)
-
-                _first_copy_prepared.set()
 
             # special treatment on /boot folder
-            _mount_point = self.mount_point
-            if Path("/boot") in entry.path.parents:
-                _mount_point = self.boot_dir
+            _mount_point = (
+                self.new_bank_root
+                if not str(entry.path).startswith("/boot")
+                else self.boot_dir
+            )
 
             # prepare this entry
             if entry.nlink == 1:
-                if is_last:  # move the tmp entry to the dst
-                    entry.move_from_src(_first_copy, _mount_point=_mount_point)
+                if is_last and not skip_cleanup:  # move the tmp entry to the dst
+                    entry.move_from_src(_first_copy, dst_root=_mount_point)
                 else:  # copy from the tmp dir
-                    entry.copy_from_src(_first_copy, _mount_point=_mount_point)
+                    entry.copy_from_src(_first_copy, dst_root=_mount_point)
             else:
-                _stat.op = "link"
                 # NOTE(20220523): for regulars.txt that support hardlink group,
                 #   use inode to identify the hardlink group.
                 #   otherwise, use hash to identify the same hardlink file.
-                _identifier = entry.sha256hash
-                if entry.inode:
-                    _identifier = entry.inode
+                _identifier = entry.sha256hash if entry.inode is None else entry.inode
 
-                _dst = _mount_point / entry.path.relative_to("/")
-                _hardlink_tracker, _is_writer = hardlink_register.get_tracker(
+                _dst = entry.relative_to(_mount_point)
+                _hardlink_tracker, _is_writer = self._hardlink_register.get_tracker(
                     _identifier, _dst, entry.nlink
                 )
                 if _is_writer:
-                    entry.copy_from_src(_first_copy, _mount_point=_mount_point)
-                    _hardlink_tracker.writer_done()
+                    entry.copy_from_src(_first_copy, dst_root=_mount_point)
                 else:
-                    _src = _hardlink_tracker.subscribe()
+                    _stat.op = "link"
+                    _src = _hardlink_tracker.subscribe_no_wait()
                     _src.link_to(_dst)
 
-                if is_last:  # cleanup last entry in tmp if not needed
+                if is_last and not skip_cleanup:
                     _first_copy.unlink(missing_ok=True)
 
+            # finish up, collect stats
             _stat.elapsed = time.thread_time() - _start
             stats_list.append(_stat)
 
         return stats_list
 
-    ###### exposed API ######
     def create_standby_bank(self):
         # TODO: erase bank on create_standby_bank
         self._prepare_meta_files()  # download meta and calculate
@@ -264,3 +265,5 @@ class RebuildMode:
         self._process_regulars()
         self._process_symlinks()
         self._process_persistents()
+
+        # TODO: cleanup the the tmp folder
